modofiquei o valor monetário do relatório de desempenho para bigdecimal

+++ barbearia-backend/dominio-principal/src/main/java/com/cesarschool/barbearia/dominio/principal/profissional/relatorio/RelatorioDesempenho.java
@@ -1,16 +1,18 @@
 package com.cesarschool.barbearia.dominio.principal.profissional.relatorio;
 
+import java.math.BigDecimal; // Importar BigDecimal
+
 import lombok.Builder;
 
 public class RelatorioDesempenho {
     private final double tempoServico; 
-    private final double receitaGerada; 
+    private final BigDecimal receitaGerada; // Alterar para BigDecimal
     private final int numeroClientesAtendidos; 
     private final double avaliacaoFuncionario; 
 
     @Builder
     public RelatorioDesempenho(double tempoServico,
-                               double receitaGerada,
+                               BigDecimal receitaGerada, // Alterar para BigDecimal
                                int numeroClientesAtendidos,
                                double avaliacaoFuncionario) {
         this.tempoServico = tempoServico;
@@ -20,7 +22,7 @@ public class RelatorioDesempenho {
     }
 
     public double getTempoServico() { return tempoServico; }
-    public double getReceitaGerada() { return receitaGerada; }
+    public BigDecimal getReceitaGerada() { return receitaGerada; } // Alterar para BigDecimal
     public int getNumeroClientesAtendidos() { return numeroClientesAtendidos; }
     public double getAvaliacaoFuncionario() { return avaliacaoFuncionario; }
 }
